refactor: created lib ethereum
