+++ libs/ethereum/index.ts
@@ -1,0 +1,3 @@
+import Web3Websocket from "./web3Websocket";
+
+export { Web3Websocket };
