+++ libs/ethereum/web3Websocket.ts
@@ -1,6 +1,6 @@
 import Web3 from "web3";
 import { WebsocketProvider } from "web3-core";
-import logger from "./logger";
+import logger from "../utils/logger";
 
 const WEB3_PROVIDER_OPTIONS = {
   clientConfig: {
