+++ libs/utils/index.ts
@@ -1,4 +1,3 @@
 import parentLogger from "./logger";
-import Web3Websocket from "./web3Websocket";
 
-export { parentLogger, Web3Websocket };
+export { parentLogger };
